requirements.txt
fastapi
uvicorn[standard]
transformers
PyJWT==2.9.0
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0